adding CommentPagingQuery

+++ src/test/java/com/example/noticeboardapi/comment/repository/CommentCommandRepositoryTest.java
@@ -2,6 +2,7 @@ package com.example.noticeboardapi.comment.repository;
 
 import com.example.noticeboardapi.comment.entity.Comment;
 import org.jooq.DSLContext;
+import org.jooq.types.ULong;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,11 +24,11 @@ public class CommentCommandRepositoryTest {
     void commentDeleteTest(Long commentNo) {
         dslContext.update(COMMENT)
                 .set(COMMENT.TEXT, "삭제된 댓글입니다.")
-                .where(COMMENT.COMMENT_ID.eq(commentNo))
+                .where(COMMENT.COMMENT_ID.eq(ULong.valueOf(commentNo)))
                 .execute();
 
         Comment comment = dslContext.selectFrom(COMMENT)
-                .where(COMMENT.COMMENT_ID.eq(commentNo))
+                .where(COMMENT.COMMENT_ID.eq(ULong.valueOf(commentNo)))
                 .fetchOneInto(Comment.class);
 
         assertThat(comment.getText()).isEqualTo("삭제된 댓글입니다.");
